NotificationService/requirements.txt
grpcio
grpcio-tools
protobuf
psycopg2-binary
SQLAlchemy
fastapi
uvicorn
websockets